Move QuickActionProcessor to a separate class
